refactor auth context into services and hooks
